Add ControllerButtonEvent
