+++ include/controller_axis_event.h
@@ -31,86 +31,10 @@
 
 #include "base_event.h"
 #include "centurion_api.h"
+#include "game_controller.h"
 
 namespace centurion {
 namespace event {
-
-/**
- * The GameControllerAxis enum class mirrors the values of the
- * SDL_GameControllerAxis enum.
- *
- * @see SDL_GameControllerAxis
- * @since 4.0.0
- */
-enum class GameControllerAxis {
-  Invalid = SDL_CONTROLLER_AXIS_INVALID,
-  LeftX = SDL_CONTROLLER_AXIS_LEFTX,
-  LeftY = SDL_CONTROLLER_AXIS_LEFTY,
-  RightX = SDL_CONTROLLER_AXIS_RIGHTX,
-  RightY = SDL_CONTROLLER_AXIS_RIGHTY,
-  TriggerLeft = SDL_CONTROLLER_AXIS_TRIGGERLEFT,
-  TriggerRight = SDL_CONTROLLER_AXIS_TRIGGERRIGHT,
-  Max = SDL_CONTROLLER_AXIS_MAX
-};
-
-/**
- * Indicates whether or not the two game controller axis values are the same.
- *
- * @param axis the lhs Centurion game controller axis value.
- * @param sdlAxis the rhs SDL game controller axis value.
- * @return true if the game controller axis values are the same; false
- * otherwise.
- * @since 4.0.0
- */
-CENTURION_NODISCARD
-CENTURION_API bool operator==(GameControllerAxis axis,
-                              SDL_GameControllerAxis sdlAxis) noexcept;
-
-/**
- * Indicates whether or not the two game controller axis values are the same.
- *
- * @param sdlAxis the lhs SDL game controller axis value.
- * @param axis the rhs Centurion game controller axis value.
- * @return true if the game controller axis values are the same; false
- * otherwise.
- * @since 4.0.0
- */
-CENTURION_NODISCARD
-CENTURION_API bool operator==(SDL_GameControllerAxis sdlAxis,
-                              GameControllerAxis axis) noexcept;
-
-/**
- * Indicates whether or not the two game controller axis values aren't the same.
- *
- * @param axis the lhs Centurion game controller axis value.
- * @param sdlAxis the rhs SDL game controller axis value.
- * @return true if the game controller axis values aren't the same; false
- * otherwise.
- * @since 4.0.0
- */
-CENTURION_NODISCARD
-CENTURION_API bool operator!=(GameControllerAxis axis,
-                              SDL_GameControllerAxis sdlAxis) noexcept;
-
-/**
- * Indicates whether or not the two game controller axis values aren't the same.
- *
- * @param sdlAxis the lhs SDL game controller axis value.
- * @param axis the rhs Centurion game controller axis value.
- * @return true if the game controller axis values aren't the same; false
- * otherwise.
- * @since 4.0.0
- */
-CENTURION_NODISCARD
-CENTURION_API bool operator!=(SDL_GameControllerAxis sdlAxis,
-                              GameControllerAxis axis) noexcept;
-
-/**
- * A type alias for SDL_JoystickID.
- *
- * @since 4.0.0
- */
-using JoystickID = SDL_JoystickID;
 
 /**
  * The ControllerAxisEvent represents an event triggered by game controller
