+++ test/CMakeLists.txt
@@ -36,8 +36,9 @@ set(TEST_FILES
         unittests/base_event_test.cpp
         unittests/audio_device_event_test.cpp
         unittests/controller_axis_event_test.cpp
+        unittests/controller_button_event_test.cpp
         unittests/game_controller_button_test.cpp
-        )
+        unittests/game_controller_axis_test.cpp)
 
 if (NOT DEFINED ENV{TRAVIS})
     set(TEST_FILES
